msvcでビルドエラーと警告を修正

+++ src/client/client_ws.cc
@@ -46,10 +46,10 @@ void Internal::messageThreadMain(std::shared_ptr<Internal::ClientData> data,
                             closed = true;
                             break;
                         } else if (meta && meta->offset > buf_s.size()) {
-                            buf_s.append(meta->offset - buf_s.size(), '\0');
+                            buf_s.append(static_cast<std::size_t>(meta->offset - buf_s.size()), '\0');
                             buf_s.append(buffer, rlen);
                         } else if (meta && meta->offset < buf_s.size()) {
-                            buf_s.replace(meta->offset, rlen, buffer, rlen);
+                            buf_s.replace(static_cast<std::size_t>(meta->offset), rlen, buffer, rlen);
                         } else {
                             buf_s.append(buffer, rlen);
                         }
